refactor(types): Convert all interface declarations to type aliases in centralized types
This commit completes the type system standardization by converting all interface
declarations to type aliases across the centralized type modules, ensuring
consistent type definition patterns throughout the codebase.

Key changes:
- Converted 7 interfaces to type aliases in src/lib/types/utilities.ts
  (LogEntry, EnhancementOptions, EnhancementResult, ConflictDetectionPlugin,
  RetryOptions, PromptRedactionOptions, EnvVarValidationResult)
- Converted EnhancedValidationResult interface to type in src/lib/types/tools.ts
- Converted TypedEventEmitter interface to type in src/lib/types/common.ts
- Updated TypedEventEmitter return types for proper type compatibility
- Fixed all import paths to use relative imports instead of $lib aliases
- Maintained full TypeScript compilation without errors

This standardization eliminates interface/type mixing and establishes type aliases
as the single pattern for all type definitions in the centralized type system,
improving consistency and maintainability.
