+++ src/lib/utils/providerUtils.ts
@@ -5,7 +5,8 @@
 import { AIProviderFactory } from "../core/factory.js";
 import { logger } from "./logger.js";
 import type { UnknownRecord } from "../types/common.js";
-import type { ProviderError } from "../types/index.js";
+import type { ProviderError } from "../types/providers.js";
+import type { EnvVarValidationResult } from "../types/utilities.js";
 import { AIProviderName } from "../constants/enums.js";
 import { ProviderHealthChecker } from "./providerHealth.js";
 import {
@@ -165,16 +166,6 @@ async function isProviderAvailable(providerName: string): Promise<boolean> {
   } catch {
     return false;
   }
-}
-
-/**
- * Validation results for environment variables
- */
-export interface EnvVarValidationResult {
-  isValid: boolean;
-  missingVars: string[];
-  invalidVars: string[];
-  warnings: string[];
 }
 
 /**
